Conflicts:
	src/applications/phriction/controller/PhrictionDocumentController.php

+++ src/applications/conpherence/controller/ConpherenceUpdateController.php
@@ -31,6 +31,7 @@ final class ConpherenceUpdateController
       ->executeOne();
 
     $action = $request->getStr('action', ConpherenceUpdateActions::METADATA);
+
     $latest_transaction_id = null;
     $response_mode = $request->isAjax() ? 'ajax' : 'redirect';
     $error_view = null;
@@ -38,7 +39,7 @@ final class ConpherenceUpdateController
     $errors = array();
     $delete_draft = false;
     $xactions = array();
-    if ($request->isFormPost()) {
+    if ($request->isFormPost() || ($action == ConpherenceUpdateActions::LOAD)) {
       $editor = id(new ConpherenceEditor())
         ->setContinueOnNoEffect($request->isContinueRequest())
         ->setContentSourceFromRequest($request)
@@ -114,24 +115,32 @@ final class ConpherenceUpdateController
               'That was a non-update. Try cancel.');
           }
           break;
+        case ConpherenceUpdateActions::LOAD:
+          $updated = false;
+          $response_mode = 'ajax';
+          break;
         default:
           throw new Exception('Unknown action: '.$action);
           break;
       }
-      if ($xactions) {
-        try {
-          $xactions = $editor->applyTransactions($conpherence, $xactions);
-          if ($delete_draft) {
-            $draft = PhabricatorDraft::newFromUserAndKey(
-              $user,
-              $conpherence->getPHID());
-            $draft->delete();
+
+      if ($xactions || ($action == ConpherenceUpdateActions::LOAD)) {
+        if ($xactions) {
+          try {
+            $xactions = $editor->applyTransactions($conpherence, $xactions);
+            if ($delete_draft) {
+              $draft = PhabricatorDraft::newFromUserAndKey(
+                $user,
+                $conpherence->getPHID());
+              $draft->delete();
+            }
+          } catch (PhabricatorApplicationTransactionNoEffectException $ex) {
+            return id(new PhabricatorApplicationTransactionNoEffectResponse())
+              ->setCancelURI($this->getApplicationURI($conpherence_id.'/'))
+              ->setException($ex);
           }
-        } catch (PhabricatorApplicationTransactionNoEffectException $ex) {
-          return id(new PhabricatorApplicationTransactionNoEffectResponse())
-            ->setCancelURI($this->getApplicationURI($conpherence_id.'/'))
-            ->setException($ex);
         }
+
         switch ($response_mode) {
           case 'ajax':
             $latest_transaction_id = $request->getInt('latest_transaction_id');
@@ -197,12 +206,15 @@ final class ConpherenceUpdateController
         id(new AphrontFormTokenizerControl())
         ->setName('add_person')
         ->setUser($user)
-        ->setDatasource('/typeahead/common/users/'));
+        ->setDatasource(new PhabricatorPeopleDatasource()));
 
     require_celerity_resource('conpherence-update-css');
     return id(new AphrontDialogView())
       ->setTitle(pht('Add Participants'))
       ->addHiddenInput('action', 'add_person')
+      ->addHiddenInput(
+        'latest_transaction_id',
+        $request->getInt('latest_transaction_id'))
       ->appendChild($form);
     }
 
@@ -232,8 +244,11 @@ final class ConpherenceUpdateController
     return id(new AphrontDialogView())
       ->setTitle(pht('Remove Participants'))
       ->addHiddenInput('action', 'remove_person')
-      ->addHiddenInput('__continue__', true)
       ->addHiddenInput('remove_person', $remove_person)
+      ->addHiddenInput(
+        'latest_transaction_id',
+        $request->getInt('latest_transaction_id'))
+      ->addHiddenInput('__continue__', true)
       ->appendChild($body);
   }
 
@@ -241,6 +256,7 @@ final class ConpherenceUpdateController
     ConpherenceThread $conpherence,
     $error_view) {
 
+    $request = $this->getRequest();
     $form = id(new PHUIFormLayoutView())
       ->appendChild($error_view)
       ->appendChild(
@@ -253,6 +269,9 @@ final class ConpherenceUpdateController
     return id(new AphrontDialogView())
       ->setTitle(pht('Update Conpherence'))
       ->addHiddenInput('action', 'metadata')
+      ->addHiddenInput(
+        'latest_transaction_id',
+        $request->getInt('latest_transaction_id'))
       ->addHiddenInput('__continue__', true)
       ->appendChild($form);
   }
@@ -266,6 +285,7 @@ final class ConpherenceUpdateController
     $need_transactions = false;
     switch ($action) {
       case ConpherenceUpdateActions::METADATA:
+      case ConpherenceUpdateActions::LOAD:
         $need_transactions = true;
         break;
       case ConpherenceUpdateActions::MESSAGE:
